requirements.txt
pandas
openai
matplotlib
tqdm
scikit-learn
python-dotenv
transformers
torch
xgboost